refactor: remove $lib/repo.ts
combine $lib/repo.ts and $lib/stores/repo.ts together
